requirements.txt
face_recognition==1.3.0
numpy==1.24.4
requests==2.31.0
Pillow==10.3.0
fastapi==0.111.0
uvicorn==0.30.1
jinja2==3.1.4
python-multipart==0.0.9
